MES-1013 - Refactor Service Layer for Business Logic Adaptation
